requirements.txt
Django==5.2.5
djangorestframework==3.16.1
numpy==2.3.2
python-dotenv==1.1.1
scikit_learn==1.7.1
sentence_transformers==5.1.0
spotipy==2.25.1
